Add support for salting MD5 values.
This resolves #1. Since the plugin stores the hashes as
hidden values in the form we should salt the hashes using a
secret token. This commit adds a configuration value to
store the salt, and then salts the hashes using this. This
insures that spammers cannot easily guess the solution by
comparing values in the question and hashing them blindly.

See the advanced usage section here for more information.

http://textcaptcha.com/how_it_works